+++ packages/core/src/index.ts
@@ -8,6 +8,7 @@ export * from './config/index';
 export * from './plugin/index';
 export * from './event-bus/index';
 export * from './process-context/index';
+export * from './picker-context/index'
 export * from './service/index';
 export * from './i18n/index';
 export * from './worker/index';